backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legalease-ai-backend"
version = "0.1.0"
description = "LegalEase AI Backend - Legal document analysis platform"
authors = [{name = "LegalEase AI Team"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "openai>=1.3.7",
    "langchain>=0.0.340",
    "langchain-openai>=0.0.2",
    "pinecone-client>=2.2.4",
    "tiktoken>=0.5.2",
    "PyPDF2>=3.0.1",
    "pdfplumber>=0.10.3",
    "pytesseract>=0.3.10",
    "Pillow>=10.1.0",
    "websockets>=12.0",
    "httpx>=0.25.2",
    "python-magic>=0.4.27",
    "python-magic-bin>=0.4.14; platform_system=='Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "psutil>=5.9.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true